+++ Ionic5Angular9Apps/test-angular-plugin-lib/src/app/app-routing.module.ts
@@ -1,16 +1,12 @@
 import { NgModule } from '@angular/core';
-
 import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
-
 
 const routes: Routes = [
   { path: '', redirectTo: 'home', pathMatch: 'full' },
   { path: 'home', loadChildren: () => import('./home/home.module').then( m => m.HomePageModule)},
-  {
-    path: 'embedded',
-    loadChildren: () => import('./embedded/embedded.module').then( m => m.EmbeddedPageModule)
-  },
+  { path: 'embedded', loadChildren: './embedded/embedded.module#EmbeddedPageModule'},
 ];
+
 @NgModule({
   imports: [
     RouterModule.forRoot(routes, { preloadingStrategy: PreloadAllModules })