+++ Ionic5Angular9Apps/test-angular-plugin-lib/src/app/app.component.ts
@@ -22,7 +22,6 @@ export class AppComponent {
     this.platform.ready().then(() => {
       this.statusBar.styleDefault();
       this.splashScreen.hide();
-      console.log("this.platform ",this.platform.is("electron"));
     });
   }
 }