+++ Ionic5Angular9Apps/test-angular-plugin-lib/src/app/embedded/embedded.module.ts
@@ -2,17 +2,21 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
 import { FormsModule } from '@angular/forms';
+import { RouterModule } from '@angular/router';
 
 import { EmbeddedPage } from './embedded.page';
-import { EmbeddedPageRoutingModule } from './embedded-routing.module';
-
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     IonicModule,
-    EmbeddedPageRoutingModule
+    RouterModule.forChild([
+      {
+        path: '',
+        component: EmbeddedPage
+      }
+    ])
   ],
   declarations: [EmbeddedPage]
 })