+++ Ionic5ReactApps/test-react-component-lib/src/utils/utils.ts
@@ -1,0 +1,5 @@
+export const getBackButtonText = (text:string): string => {
+    const win = window as any;
+    const mode = win && win.Ionic && win.Ionic.mode;
+    return mode === 'ios' ? text : '';
+  }